django_weather_app/requirements.txt
django==5.0.7
psycopg2-binary>=2.9.6,<2.10
djangorestframework==3.15.2
Markdown==3.6
requests==2.32.3